Decrypt SSH key before use
The SSH key is now encrypted with a RSA public key before being sent to
the terminal service.  The terminal service needs to be configured with
the corresponding RSA private key.  If it is, the SSH key will be
decrypted and used.  If not a "failure to connect" message will be
shown.