Rework configuration system
Previously, we had a poorly documented mix of environment variables and
a JSON config file.  Now as much as possible has been moved to
environment variables configured through dotenv.  The only configuration
that has not been moved to environment variables are the default
arguments to provide to the SSH command.